+++ Services/VaultsService.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using keepr.Models;
 using keepr.Repositories;
 
@@ -23,21 +24,48 @@ namespace keepr.Services
 
     internal IEnumerable<Vault> Get(string profileId)
     {
-      return _repo.Get(profileId);
+      return _repo.Get(profileId).ToList().FindAll(vault => vault.CreatorId == profileId || !vault.IsPrivate);
     }
 
-    internal string Delete(int id)
+    internal string Delete(int id, Profile userInfo)
     {
+
+      Vault returnedVault = _repo.GetOne(id);
+      if (returnedVault == null)
+      {
+        throw new Exception("Does not exist");
+      }
+      if (returnedVault.CreatorId != userInfo.Id)
+      {
+        throw new Exception("Access Denied");
+      }
       if (_repo.Delete(id))
       {
         return "Deleted!";
       }
-      return "Unable to be deleted";
+      else
+      {
+        return "Failed";
+      }
+
     }
 
-    internal Vault GetOne(int id)
+    internal Vault GetOne(int id, Profile userInfo)
     {
-      return _repo.GetOne(id);
+
+      Vault returnedVault = _repo.GetOne(id);
+      if (returnedVault == null)
+      {
+        throw new Exception("Does not exist");
+      }
+      else if (!returnedVault.IsPrivate || returnedVault.CreatorId == userInfo.Id)
+      {
+        return returnedVault;
+      }
+      else
+      {
+        throw new Exception("Access Denied");
+      }
     }
 
     internal Vault Edit(Vault editedVault, Profile userInfo)
@@ -53,6 +81,12 @@ namespace keepr.Services
       }
       _repo.Edit(editedVault);
       return _repo.GetOne(editedVault.Id);
+    }
+
+    internal IEnumerable<Vault> GetVaultsByProfile(string profileId, string userId)
+    {
+      return _repo.GetVaultsByProfile(profileId).ToList().FindAll(vault => (vault.CreatorId == profileId && !vault.IsPrivate) || vault.CreatorId == userId);
+
     }
   }
 }
